+++ lib/src/core/bean/utils/dart_ddi_utils.dart
@@ -1,7 +1,6 @@
 import 'package:dart_ddi/dart_ddi.dart';
 import 'package:dart_ddi/src/data/factory_clazz.dart';
 import 'package:dart_ddi/src/exception/bean_destroyed.dart';
-import 'package:dart_ddi/src/exception/duplicated_bean.dart';
 import 'package:dart_ddi/src/typedef/typedef.dart';
 
 final class DartDDIUtils {
@@ -23,16 +22,6 @@ final class DartDDIUtils {
     final PostConstruct clazz = await register;
 
     return clazz.onPostConstruct();
-  }
-
-  static void validateDuplicated(Object effectiveQualifierName, bool debug) {
-    if (debug) {
-      // ignore: avoid_print
-      print(
-          'Is already registered a instance with Type ${effectiveQualifierName.toString()}');
-    } else {
-      throw DuplicatedBeanException(effectiveQualifierName.toString());
-    }
   }
 
   static BeanT getSingleton<BeanT extends Object>(