+++ lib/src/core/bean/utils/dispose_utils.dart
@@ -3,7 +3,8 @@ import 'package:dart_ddi/src/data/factory_clazz.dart';
 
 final class DisposeUtils {
   /// Dispose only clean the class Instance
-  static Future<void> disposeBean<BeanT>(FactoryClazz<BeanT> factoryClazz) {
+  static Future<void> disposeBean<BeanT extends Object>(
+      FactoryClazz<BeanT> factoryClazz) {
     if (factoryClazz.interceptors case final inter? when inter.isNotEmpty) {
       // Call aroundDispose before reset the clazzInstance
       // Should call interceptors even if the instance is null
@@ -16,13 +17,14 @@ final class DisposeUtils {
       return _runFutureOrPreDispose<BeanT>(factoryClazz, clazz);
     }
 
-    _disposeChildren(factoryClazz);
+    _disposeChildren(factoryClazz.children);
+    factoryClazz.clazzInstance = null;
     return Future.value();
   }
 
-  static Future<void> _runFutureOrPreDispose<BeanT>(
+  static Future<void> _runFutureOrPreDispose<BeanT extends Object>(
       FactoryClazz<BeanT> factoryClazz, PreDispose clazz) async {
-    disposeChildrenAsync<BeanT>(factoryClazz);
+    disposeChildrenAsync<BeanT>(factoryClazz.children);
 
     await clazz.onPreDispose();
 
@@ -31,22 +33,18 @@ final class DisposeUtils {
     return Future.value();
   }
 
-  static void _disposeChildren<BeanT>(FactoryClazz<BeanT> factoryClazz) {
-    if (factoryClazz.children case final List<Object> children?
-        when children.isNotEmpty) {
-      for (final Object child in children) {
+  static void _disposeChildren<BeanT extends Object>(Set<Object>? children) {
+    if (children?.isNotEmpty ?? false) {
+      for (final Object child in children!) {
         ddi.dispose(qualifier: child);
       }
     }
-
-    factoryClazz.clazzInstance = null;
   }
 
-  static Future<void> disposeChildrenAsync<BeanT>(
-      FactoryClazz<BeanT> factoryClazz) async {
-    if (factoryClazz.children case final List<Object> children?
-        when children.isNotEmpty) {
-      for (final Object child in children) {
+  static Future<void> disposeChildrenAsync<BeanT extends Object>(
+      Set<Object>? children) async {
+    if (children?.isNotEmpty ?? false) {
+      for (final Object child in children!) {
         await ddi.dispose(qualifier: child);
       }
     }