+++ lib/src/mixin/ddi_inject_mixin.dart
@@ -4,7 +4,7 @@ import 'package:dart_ddi/dart_ddi.dart';
 ///
 /// Example:
 /// ```dart
-/// class MyController with DDIInject {
+/// class MyController with DDIInject<MyService> {
 ///
 ///   void businessLogic() {
 ///     instance.runSomething();
@@ -13,6 +13,21 @@ import 'package:dart_ddi/dart_ddi.dart';
 /// ```
 mixin DDIInject<InjectType extends Object> {
   final InjectType instance = ddi.get<InjectType>();
+}
+
+/// Helper to make easy to Inject a Component instance
+///
+/// Example:
+/// ```dart
+/// class MyController with DDIComponentInject<MyComponent, MyModule> {
+///
+///   void businessLogic() {
+///     instance.runSomething();
+///   }
+/// }
+/// ```
+mixin DDIComponentInject<ComponentT extends Object, ModuleT extends DDIModule> {
+  final ComponentT instance = ddi.getComponent<ComponentT>(module: ModuleT);
 }
 
 /// Helper to make easy to Inject one instance