+++ lib/src/mixin/ddi_module_mixin.dart
@@ -28,7 +28,7 @@ mixin DDIModule implements PostConstruct {
     _internalQualifier = value;
   }
 
-  List<Object> get children => ddi.getChildren(qualifier: moduleQualifier);
+  Set<Object> get children => ddi.getChildren(qualifier: moduleQualifier);
 
   /// Registers an instance as a Singleton.
   ///
@@ -39,6 +39,7 @@ mixin DDIModule implements PostConstruct {
   /// - `interceptor`: Optional interceptor to customize the creation, get, dispose or remove behavior.
   /// - `registerIf`: Optional function to conditionally register the instance.
   /// - `destroyable`: Optional parameter to make the instance indestructible.
+  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to vinculate multiple classes under a single parent module.
   Future<void> registerSingleton<BeanT extends Object>(
     BeanRegister<BeanT> clazzRegister, {
     Object? qualifier,
@@ -47,7 +48,7 @@ mixin DDIModule implements PostConstruct {
     ListDDIInterceptor<BeanT>? interceptors,
     FutureOrBoolCallback? registerIf,
     bool destroyable = true,
-    List<Object>? children,
+    Set<Object>? children,
   }) {
     ddi.addChildModules(child: qualifier ?? BeanT, qualifier: moduleQualifier);
 
@@ -72,7 +73,7 @@ mixin DDIModule implements PostConstruct {
   /// - `interceptor`: Optional interceptor to customize the creation, get, dispose or remove behavior.
   /// - `registerIf`: Optional function to conditionally register the instance.
   /// - `destroyable`: Optional parameter to make the instance indestructible.
-  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to register multiple classes under a single parent module.
+  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to vinculate multiple classes under a single parent module.
   Future<void> registerApplication<BeanT extends Object>(
     BeanRegister<BeanT> clazzRegister, {
     Object? qualifier,
@@ -81,7 +82,7 @@ mixin DDIModule implements PostConstruct {
     ListDDIInterceptor<BeanT>? interceptors,
     FutureOrBoolCallback? registerIf,
     bool destroyable = true,
-    List<Object>? children,
+    Set<Object>? children,
   }) {
     ddi.addChildModules(child: qualifier ?? BeanT, qualifier: moduleQualifier);
 
@@ -106,7 +107,7 @@ mixin DDIModule implements PostConstruct {
   /// - `interceptor`: Optional interceptor to customize the creation, get, dispose or remove behavior.
   /// - `registerIf`: Optional function to conditionally register the instance.
   /// - `destroyable`: Optional parameter to make the instance indestructible.
-  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to register multiple classes under a single parent module.
+  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to vinculate multiple classes under a single parent module.
   Future<void> registerSession<BeanT extends Object>(
     BeanRegister<BeanT> clazzRegister, {
     Object? qualifier,
@@ -115,7 +116,7 @@ mixin DDIModule implements PostConstruct {
     ListDDIInterceptor<BeanT>? interceptors,
     FutureOrBoolCallback? registerIf,
     bool destroyable = true,
-    List<Object>? children,
+    Set<Object>? children,
   }) {
     ddi.addChildModules(child: qualifier ?? BeanT, qualifier: moduleQualifier);
 
@@ -140,7 +141,7 @@ mixin DDIModule implements PostConstruct {
   /// - `interceptor`: Optional interceptor to customize the creation, get, dispose or remove behavior.
   /// - `registerIf`: Optional function to conditionally register the instance.
   /// - `destroyable`: Optional parameter to make the instance indestructible.
-  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to register multiple classes under a single parent module.
+  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to vinculate multiple classes under a single parent module.
   Future<void> registerDependent<BeanT extends Object>(
     BeanRegister<BeanT> clazzRegister, {
     Object? qualifier,
@@ -149,7 +150,7 @@ mixin DDIModule implements PostConstruct {
     ListDDIInterceptor<BeanT>? interceptors,
     FutureOrBoolCallback? registerIf,
     bool destroyable = true,
-    List<Object>? children,
+    Set<Object>? children,
   }) {
     ddi.addChildModules(child: qualifier ?? BeanT, qualifier: moduleQualifier);
 
@@ -174,7 +175,7 @@ mixin DDIModule implements PostConstruct {
   /// - `interceptor`: Optional interceptor to customize the creation, get, dispose or remove behavior.
   /// - `registerIf`: Optional function to conditionally register the instance.
   /// - `destroyable`: Optional parameter to make the instance indestructible.
-  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to register multiple classes under a single parent module.
+  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to vinculate multiple classes under a single parent module.
   Future<void> registerObject<BeanT extends Object>(
     BeanT register, {
     Object? qualifier,
@@ -183,12 +184,57 @@ mixin DDIModule implements PostConstruct {
     ListDDIInterceptor<BeanT>? interceptors,
     FutureOrBoolCallback? registerIf,
     bool destroyable = true,
-    List<Object>? children,
+    Set<Object>? children,
   }) {
     ddi.addChildModules(child: qualifier ?? BeanT, qualifier: moduleQualifier);
 
     return ddi.registerObject<BeanT>(
       register,
+      qualifier: qualifier,
+      postConstruct: postConstruct,
+      decorators: decorators,
+      interceptors: interceptors,
+      destroyable: destroyable,
+      registerIf: registerIf,
+      children: children,
+    );
+  }
+
+  /// Registers an instance as a Component.
+  ///
+  /// Registering the instance with this method means that "This instance is for this module only".
+  ///
+  /// To retrieve the instance, requires to use with `ddi.getComponent`
+  ///
+  /// Example:
+  /// ```dart
+  /// MyWidgetComponent get myWidget => ddi.getComponent(MyModule);
+  /// ```
+  ///
+  /// Useful for Flutter apps where you want to reuse the same component or instance across specific routes or Widgets.
+  /// Additionally, you can inject and retrieve the same component in each "SubModule" or even in a "SubModule" get the parent module instance.
+  ///
+  /// - `clazzRegister`: Factory function to create the instance.
+  /// - `qualifier`: Optional qualifier name to distinguish between different instances of the same type.
+  /// - `postConstruct`: Optional function to be executed after the instance is constructed.
+  /// - `decorators`: List of decoration functions to apply to the instance.
+  /// - `interceptor`: Optional interceptor to customize the creation, get, dispose or remove behavior.
+  /// - `registerIf`: Optional function to conditionally register the instance.
+  /// - `destroyable`: Optional parameter to make the instance indestructible.
+  /// - `children`: Optional parameter, designed to receive types or qualifiers. This parameter allows you to vinculate multiple classes under a single parent module.
+  Future<void> registerComponent<BeanT extends Object>(
+    BeanRegister<BeanT> clazzRegister, {
+    Object? qualifier,
+    VoidCallback? postConstruct,
+    ListDecorator<BeanT>? decorators,
+    ListDDIInterceptor<BeanT>? interceptors,
+    FutureOrBoolCallback? registerIf,
+    bool destroyable = true,
+    Set<Object>? children,
+  }) {
+    return ddi.registerComponent<BeanT>(
+      clazzRegister: clazzRegister,
+      moduleQualifier: moduleQualifier,
       qualifier: qualifier,
       postConstruct: postConstruct,
       decorators: decorators,