+++ test/clazz_samples/component.dart
@@ -1,0 +1,4 @@
+class Component {
+  const Component(this.value);
+  final String value;
+}