+++ test/dart_ddi_test.dart
@@ -1,5 +1,3 @@
-import 'package:dart_ddi/dart_ddi.dart';
-
 import 'beans_test/add_decoratos_test.dart';
 import 'beans_test/application_future_test.dart';
 import 'beans_test/application_test.dart';
@@ -14,6 +12,7 @@ import 'beans_test/get_by_future_type_test.dart';
 import 'beans_test/get_by_type_test.dart';
 import 'beans_test/interceptor_test.dart';
 import 'beans_test/module_application_test.dart';
+import 'beans_test/module_component_test.dart';
 import 'beans_test/module_dependent_test.dart';
 import 'beans_test/module_object_test.dart';
 import 'beans_test/module_session_test.dart';
@@ -33,7 +32,6 @@ import 'event_test/timer_events_test.dart';
 import 'stream_test/stream_test.dart';
 
 void main() {
-  ddi.setDebugMode(false);
   //Basic Tests, with consists in register, get, dispose, remove
   singleton();
   application();
@@ -74,6 +72,7 @@ void main() {
   moduleDependentTest();
   moduleObjectTest();
   moduleSessionTest();
+  moduleComponentTest();
 
   //Events
   eventTest();